fix: comments and stuff
